% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\docType{methods}
\name{BoundFc,character-method}
\alias{BoundFc,character-method}
\title{the source position of this method is found correctly}
\usage{
\S4method{BoundFc}{character}(format, ...)
}
\description{
the source position of this method is found correctly
}
