% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\docType{methods}
\name{BoundFc,missing-method}
\alias{BoundFc,missing-method}
\title{the source position of this method is not found but yields a list}
\usage{
\S4method{BoundFc}{missing}(format, ...)
}
\arguments{
\item{...}{something to be passed to TimeMap}
}
\description{
the source position of this method is not found but yields a list
}
